+++ documentation/src/components/content/ExternalContent.svelte
@@ -1,33 +1,24 @@
 <script lang="ts">
 	import { isMenuOpen } from "src/lib/stores.js";
 	export let content: Content = [];
-	export let currentTitle: string | null = null;
-	export let framework: Framework;
+	export let documentTitle: string | null = null;
+	export let title: string;
 </script>
 
 <div
 	class=" border-r lg:border-r-0 dark:border-zinc-900 lg:block shrink-0 pt-20 fixed w-72 h-screen ml-4 sm:ml-8 lg:ml-12 bg-white dark:bg-black-zinc"
 	class:hidden={!$isMenuOpen}
 >
-	{#if framework}
-		<h3 class="text-lg font-medium">
-			{#if framework === "sveltekit"}
-				SvelteKit
-			{:else if framework === "express"}
-				Express
-			{:else if framework === "nextjs"}
-				Next.js
-			{/if}
-			Integration
-		</h3>
-	{/if}
+	<h3 class="text-lg font-medium">
+		{title}
+	</h3>
 	<div class="overflow-auto h-full relative overscroll-contain pr-4 mt-2">
 		{#each content as section}
 			<div class="mb-10">
 				<p class="font-medium">{section.title}</p>
 				<ul class="list-none mt-2 text-gray-500 dark:text-zinc-400">
 					{#each section.pages as page}
-						{@const isSelected = currentTitle === page.title}
+						{@const isSelected = documentTitle === page.title}
 						<li
 							class="my-1 pl-4 border-l-2"
 							class:text-main={isSelected}