+++ packages/lucia-auth/src/auth/cookie.ts
@@ -1,12 +1,12 @@
 import { Auth, Session } from "../types.js";
-import cookie from "cookie";
+import { serialize } from "../utils/cookie.js";
 
 type CreateSessionCookies = (session: Session) => string[];
 
 export const createSessionCookiesFunction = (auth: Auth) => {
 	const createSessionCookies: CreateSessionCookies = (session) => {
 		return auth.configs.sessionCookieOptions.map((option) => {
-			return cookie.serialize("auth_session", session.sessionId, {
+			return serialize("auth_session", session.sessionId, {
 				...option,
 				httpOnly: true,
 				expires: new Date(session.idlePeriodExpires),
@@ -23,7 +23,7 @@ export const createBlankSessionCookiesFunction = (auth: Auth) => {
 	const createBlankSessionCookies: CreateBlankSessionCookies = () => {
 		const options = [...auth.configs.sessionCookieOptions, ...auth.configs.deleteCookieOptions];
 		return options.map((option) => {
-			return cookie.serialize("auth_session", "", {
+			return serialize("auth_session", "", {
 				...option,
 				httpOnly: true,
 				maxAge: 0,