+++ packages/oauth/src/ambient.d.ts
@@ -1,0 +1,13 @@
+/// <reference types="lucia-auth" />
+declare namespace Lucia {
+	export type UserAttributes = {
+		username?: string
+	}
+	export type Auth = {
+		configs: {
+			transformUserData: () => {
+				userId: string
+			}
+		}
+	};
+}